backend/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
psycopg2-binary
python-dotenv
